backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
blis==0.7.9
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
cymem==2.0.7
dataclasses-json==0.5.8
dnspython==2.3.0
fastapi==0.97.0
fr-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.5.0/fr_core_news_sm-3.5.0-py3-none-any.whl
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
langchain==0.0.207
langchainplus-sdk==0.0.16
langcodes==3.3.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pathy==0.10.2
pluggy==1.2.0
preshed==3.0.8
pydantic==1.10.9
pymongo==4.3.3
pytest==7.4.0
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
smart-open==6.3.0
sniffio==1.3.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.16
srsly==2.4.6
starlette==0.27.0
tenacity==8.2.2
thinc==8.1.10
tiktoken==0.4.0
tqdm==4.65.0
typer==0.7.0
typing-inspect==0.9.0
typing_extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0
wasabi==1.1.2
yarl==1.9.2
youtube-transcript-api==0.6.1
